Update rr to use rdata instead of a simple test string
This means some other changes. First, rdata can not be compared, so operator== was added and a virtual equals was used to virtualize the operator, this means adding equals to all derived classes

Second, I needed a named constructor for rdata that builds a derived classes returning a pointer to Rdata.

Thrid, As abstract base classes can not be instantiated, you can not use a reference on Rr for Rdata, so a pointer has been added. Then Rr needs a destructor to take care of the pointer.